Added functionality to reuse the last used path per file type.
When a project, image or animation file is loaded or saved, the next
time the user will be prompted to load or save the file in the same
directory.

Adittionally for the osx-bundle changed the current working directory to
the directory where the app is located. I wonder if this is a good idea.
:3