Added d&d for dependent of the sender and removed unnecessary stuff.
The central widget doesn't need to receive drop events, the main window
can take care of those. Only the two lateral widgets need special
handling when something is dropped on them. For example I only want the
image dropped on a particular widget to be loaded/selected on that
widget.